Сделал задание 7 модуля
